refactor(MangaUpdates): move search parser out of shared
