+++ src/MangaUpdates/Implementations/Shared/parser/main.ts
@@ -1,3 +1,2 @@
 export * as manga from "./manga";
-export * as search from "./search";
 export * as session from "./session";
